pyproject.toml
# PhoenixDT - Industrial Digital Twin

[build-system]
requires = ["setuptools>=70.0", "wheel", "setuptools_scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "phoenix-dt"
dynamic = ["version"]
description = "Industrial digital twin with motor simulation"
authors = [
    {name = "Ankit Karki", email = "karkiankit101@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Monitoring",
]
keywords = [
    "digital-twin", 
    "industrial-ai", 
    "motor-simulation", 
    "predictive-maintenance", 
    "anomaly-detection",
    "industrial-automation"
]
dependencies = [
    # Core computing
    "numpy>=1.26.0",
    "scipy>=1.13.0", 
    
    # Web framework
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    "pydantic>=2.8.0",
    
    # CLI and UI
    "click>=8.0.0",
    "rich>=13.8.0",
    
    # Utilities
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0", 
    "pytest-cov>=5.0.0",
    
    # Code quality
    "black>=24.8.0",
    "isort>=5.13.0", 
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "pre-commit>=3.8.0",
]
all = [
    "phoenix-dt[dev]",
]

[project.urls]
Homepage = "https://github.com/Ankit-x1/HealingSystem"
Repository = "https://github.com/Ankit-x1/HealingSystem"
Documentation = "https://github.com/Ankit-x1/HealingSystem/docs"
"Bug Tracker" = "https://github.com/Ankit-x1/HealingSystem/issues"

[project.scripts]
phoenixdt = "phoenixdt.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

# [tool.setuptools_scm]
# write_to = "src/phoenixdt/_version.py"

[tool.black]
line-length = 88
target-version = ['py312', 'py313']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["phoenixdt"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "C901"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "--cov=phoenixdt",
    "--cov-report=html", 
    "--cov-report=term-missing",
    "-v"
]